fix: fix theme provider into separate files
