+++ src/components/ThemeContext/ThemeContext.tsx
@@ -1,20 +1,6 @@
-import { createContext, useState, useEffect, ReactNode } from 'react';
-import { Theme, ThemeContextProps } from '../../types/types';
+import { createContext } from 'react';
+import { ThemeContextProps } from '../../types/types';
 
 export const ThemeContext = createContext<ThemeContextProps | undefined>(
   undefined
 );
-
-export const ThemeProvider = ({ children }: { children: ReactNode }) => {
-  const [theme, setTheme] = useState<Theme>('light');
-
-  useEffect(() => {
-    document.documentElement.setAttribute('data-theme', theme);
-  }, [theme]);
-
-  return (
-    <ThemeContext.Provider value={{ theme, setTheme }}>
-      {children}
-    </ThemeContext.Provider>
-  );
-};
