+++ src/components/ThemeContext/ThemeProvider.tsx
@@ -1,0 +1,17 @@
+import { useState, useEffect, ReactNode } from 'react';
+import { Theme } from '../../types/types';
+import { ThemeContext } from './ThemeContext';
+
+export const ThemeProvider = ({ children }: { children: ReactNode }) => {
+  const [theme, setTheme] = useState<Theme>('light');
+
+  useEffect(() => {
+    document.documentElement.setAttribute('data-theme', theme);
+  }, [theme]);
+
+  return (
+    <ThemeContext.Provider value={{ theme, setTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
